+++ src/kernel/src/memory/paging/table.rs
@@ -8,7 +8,7 @@ use super::{Entry, ENTRY_COUNT};
 pub const P4: *mut Table<Level4> = 0xffffffff_fffff000 as *mut _;
 
 pub struct Table<L: TableLevel> {
-    entries: [Entry; ENTRY_COUNT],
+    entries: [Entry; ENTRY_COUNT as usize],
     level: PhantomData<L>,
 }
 
@@ -27,30 +27,30 @@ impl<L> Table<L>
 where
     L: HierarchicalLevel,
 {
-    fn next_table_address(&self, index: usize) -> Option<usize> {
-        let entry_flags = self[index].flags();
+    fn next_table_address(&self, index: u64) -> Option<u64> {
+        let entry_flags = self[index as usize].flags();
         if entry_flags.contains(EntryFlags::PRESENT) && !entry_flags.contains(EntryFlags::HUGE_PAGE)
         {
-            let table_address = self as *const _ as usize;
+            let table_address = self as *const _ as u64;
             Some((table_address << 9) | (index << 12))
         } else {
             None
         }
     }
 
-    pub fn next_table(&self, index: usize) -> Option<&Table<L::NextLevel>> {
+    pub fn next_table(&self, index: u64) -> Option<&Table<L::NextLevel>> {
         self.next_table_address(index)
             .map(|address| unsafe { &*(address as *const _) })
     }
 
-    pub fn next_table_mut(&mut self, index: usize) -> Option<&mut Table<L::NextLevel>> {
+    pub fn next_table_mut(&mut self, index: u64) -> Option<&mut Table<L::NextLevel>> {
         self.next_table_address(index)
             .map(|address| unsafe { &mut *(address as *mut _) })
     }
 
     pub fn next_table_create<A>(
         &mut self,
-        index: usize,
+        index: u64,
         allocator: &mut A,
     ) -> &mut Table<L::NextLevel>
     where
@@ -58,11 +58,13 @@ where
     {
         if self.next_table(index).is_none() {
             assert!(
-                !self.entries[index].flags().contains(EntryFlags::HUGE_PAGE),
+                !self.entries[index as usize]
+                    .flags()
+                    .contains(EntryFlags::HUGE_PAGE),
                 "mapping code does not support huge pages"
             );
             let frame = allocator.allocate_frame().expect("no frames available");
-            self.entries[index].set(frame, EntryFlags::PRESENT | EntryFlags::WRITABLE);
+            self.entries[index as usize].set(frame, EntryFlags::PRESENT | EntryFlags::WRITABLE);
             self.next_table_mut(index).unwrap().zero();
         }
         self.next_table_mut(index).unwrap()
